refactor: rename text-linkify files and update email regex patterns
