+++ src/commons/mail-message-renderer/text-message-renderer/text-linkify.ts
@@ -31,14 +31,13 @@ const DEFAULT_OPTIONS: Required<Omit<LinkifyOptions, 'autolinker'>> & {
 };
 
 const MAILTO_EMAIL_REGEX =
-	// eslint-disable-next-line max-len
-	/mailto:([\p{L}\p{N}._%+-]+@(?:[\p{L}\p{N}.-]+\.[\p{L}\p{N}]{2,}|\[[^\]\s<>]+\]))(\?[^&\s<>]+(?:&(?!gt;|lt;)[^&\s<>]+)*)?(?=&[lg]t;|[\s<>]|$)/gu;
+	/mailto:([\p{L}\p{N}._%+-]+@(?:[\p{L}\p{N}.-]+\.[\p{L}\p{N}]{2,}|\[[^\]\s<>]+\]))(\?[^\s<>]+)?/gu;
 
 const PLAIN_EMAIL_REGEX =
-	/(^|[\s>])([\p{L}\p{N}._%+-]+@(?:[\p{L}\p{N}.-]+\.[\p{L}\p{N}]{2,}|\[[^\]\s<>]+\]))/gu;
+	/(^|\s)([\p{L}\p{N}._%+-]+@(?:[\p{L}\p{N}.-]+\.[\p{L}\p{N}]{2,}|\[[^\]\s<>]+\]))/gu;
 
 const ANGLE_BRACKET_EMAIL_REGEX =
-	/(?:<|&lt;)([a-zA-Z0-9._%+-]+@(?:[a-zA-Z0-9-]+\.)+[a-zA-Z0-9-]{2,}|\[[^\]\s<>]+])(?:>|&gt;)/g;
+	/<([a-zA-Z0-9._%+-]+@(?:[a-zA-Z0-9-]+\.)+[a-zA-Z0-9-]{2,}|\[[^\]\s<>]+])>/g;
 
 function asAttrs(opts: Required<LinkifyOptions>): string {
 	const target = opts.openInNewTab ? ' target="_blank"' : '';
